requirements.txt
python-slugify==3.0.0
text-unidecode==1.2
